상품 등록 및 수정 -ing

+++ src/main/java/com/somfoot/mugpet/repository/MemberRepository.java
@@ -3,5 +3,5 @@ package com.somfoot.mugpet.repository;
 import com.somfoot.mugpet.entity.MemberInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepository extends JpaRepository<MemberInfo,Integer> {
+public interface MemberRepository extends JpaRepository<MemberInfo,Long> {
 }
